refactoring charge service to a concern
