+++ app/models/credit_card.rb
@@ -3,6 +3,8 @@ class CreditCard < ApplicationRecord
   delegate :customer, to: :account
   has_many :transactions
 
+  include Chargeable
+
   validates :spent_limit, numericality: { greater_than_or_equal_to: 0 }
   validates_presence_of(
     :account_id, :spent_limit, :number, :expiration_date, :cvv
